Moving iree/base/atomic_slist.h to iree/base/internal/.
