+++ iree/base/BUILD
@@ -87,26 +87,6 @@ cc_test(
 #===------------------------------------------------------------------------===#
 
 cc_library(
-    name = "atomic_slist",
-    srcs = ["atomic_slist.c"],
-    hdrs = ["atomic_slist.h"],
-    deps = [
-        ":core_headers",
-        ":synchronization",
-    ],
-)
-
-cc_test(
-    name = "atomic_slist_test",
-    srcs = ["atomic_slist_test.cc"],
-    deps = [
-        ":atomic_slist",
-        "//iree/testing:gtest",
-        "//iree/testing:gtest_main",
-    ],
-)
-
-cc_library(
     name = "dynamic_library",
     srcs = [
         "dynamic_library_posix.cc",
