+++ iree/base/internal/BUILD
@@ -21,6 +21,26 @@ package(
 )
 
 cc_library(
+    name = "atomic_slist",
+    srcs = ["atomic_slist.c"],
+    hdrs = ["atomic_slist.h"],
+    deps = [
+        "//iree/base:core_headers",
+        "//iree/base:synchronization",
+    ],
+)
+
+cc_test(
+    name = "atomic_slist_test",
+    srcs = ["atomic_slist_test.cc"],
+    deps = [
+        ":atomic_slist",
+        "//iree/testing:gtest",
+        "//iree/testing:gtest_main",
+    ],
+)
+
+cc_library(
     name = "atomics",
     hdrs = [
         "atomics_clang.h",
