+++ iree/base/internal/CMakeLists.txt
@@ -16,6 +16,30 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
+    atomic_slist
+  HDRS
+    "atomic_slist.h"
+  SRCS
+    "atomic_slist.c"
+  DEPS
+    iree::base::core_headers
+    iree::base::synchronization
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    atomic_slist_test
+  SRCS
+    "atomic_slist_test.cc"
+  DEPS
+    ::atomic_slist
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+iree_cc_library(
+  NAME
     atomics
   HDRS
     "atomics_clang.h"
