+++ iree/hal/local/BUILD
@@ -33,9 +33,9 @@ cc_library(
     hdrs = ["arena.h"],
     deps = [
         "//iree/base:api",
-        "//iree/base:atomic_slist",
         "//iree/base:core_headers",
         "//iree/base:synchronization",
+        "//iree/base/internal:atomic_slist",
     ],
 )
